requirements.txt
fastapi==0.103.2
uvicorn==0.22.0
SQLAlchemy==2.0.21
pydantic==1.10.12
redis==4.5.5
requests==2.31.0
httpx==0.24.1
python-dotenv==1.0.0
apscheduler==3.10.1
PyMySQL==1.1.0
